+++ src/Libs/RPC/__init__.py
@@ -1,0 +1,2 @@
+# __init__.py
+from .Phantasma import PhantasmaAPI
