+++ src/readme.md
@@ -1,16 +1,14 @@
 # PhantasmaSDK
 
-## Python Sample app
+## Python SimpleWallet Sample app
 
 This is a simple wallet sample that needs to be connected to a RPC node. By default it uses the localhost:7077/rpc endpoint, but you can switch to your own URL.
 
 To run the sample app, follow these steps:
 
-1. Overwrite the phantasma.py on the sample folder with the latest one from Output/temp/Python/Libs
+1. Install "requests" module (Run "pip install requests" on command line inside app folder)
 
-2. Install "requests" module (Run "pip install requests" on command line inside app folder)
-
-3. Run the sample app!
+2. Run the sample app!
 
 ## Python VM Samples
